requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==20.1.0
Flask-Limiter==3.5.0
Pillow==11.1.0
flask-wtf
sqlalchemy
wtforms
